add retry logic in utils
